// src/services/openai/types.ts
export interface CreditCardRecommendation {
  card_name: string;
  annual_fee: string;
  maximum_value_of_benefits: string;
  real_world_benefits: string;
  why_this_card: string;
  maximum_value_description: string;
  card_image: string;
  apply_link: string;
}

export interface RecommendationsResponse {
  recommendations: CreditCardRecommendation[];
}